refactor: move filters in separate files
