+++ filters/alloffilter.h
@@ -1,0 +1,20 @@
+#ifndef ALLOFFILTER_H
+#define ALLOFFILTER_H
+
+#include "filtercontainerfilter.h"
+
+namespace qqsfpm {
+
+class AllOfFilter : public FilterContainerFilter {
+    Q_OBJECT
+
+public:
+    using FilterContainerFilter::FilterContainerFilter;
+
+protected:
+    bool filterRow(const QModelIndex& sourceIndex, const QQmlSortFilterProxyModel& proxyModel) const override;
+};
+
+}
+
+#endif // ALLOFFILTER_H
