+++ filters/anyoffilter.h
@@ -1,0 +1,20 @@
+#ifndef ANYOFFILTER_H
+#define ANYOFFILTER_H
+
+#include "filtercontainerfilter.h"
+
+namespace qqsfpm {
+
+class AnyOfFilter : public FilterContainerFilter {
+    Q_OBJECT
+
+public:
+    using FilterContainerFilter::FilterContainerFilter;
+
+protected:
+    bool filterRow(const QModelIndex& sourceIndex, const QQmlSortFilterProxyModel& proxyModel) const override;
+};
+
+}
+
+#endif // ANYOFFILTER_H
