+++ filters/expressionfilter.cpp
@@ -1,0 +1,124 @@
+#include "expressionfilter.h"
+#include "qqmlsortfilterproxymodel.h"
+#include <QtQml>
+
+namespace qqsfpm {
+
+/*!
+    \qmltype ExpressionFilter
+    \inherits Filter
+    \inqmlmodule SortFilterProxyModel
+    \brief Filters row with a custom filtering
+
+    An ExpressionFilter is a \l Filter allowing to implement custom filtering based on a javascript expression.
+*/
+
+/*!
+    \qmlproperty expression ExpressionFilter::expression
+
+    An expression to implement custom filtering, it must evaluate to a boolean.
+    It has the same syntax has a \l {http://doc.qt.io/qt-5/qtqml-syntax-propertybinding.html} {Property Binding} except it will be evaluated for each of the source model's rows.
+    Rows that have their expression evaluating to \c true will be accepted by the model.
+    Data for each row is exposed like for a delegate of a QML View.
+
+    This expression is reevaluated for a row every time its model data changes.
+    When an external property (not \c index or in \c model) the expression depends on changes, the expression is reevaluated for every row of the source model.
+    To capture the properties the expression depends on, the expression is first executed with invalid data and each property access is detected by the QML engine.
+    This means that if a property is not accessed because of a conditional, it won't be captured and the expression won't be reevaluted when this property changes.
+
+    A workaround to this problem is to access all the properties the expressions depends unconditionally at the beggining of the expression.
+*/
+const QQmlScriptString& ExpressionFilter::expression() const
+{
+    return m_scriptString;
+}
+
+void ExpressionFilter::setExpression(const QQmlScriptString& scriptString)
+{
+    if (m_scriptString == scriptString)
+        return;
+
+    m_scriptString = scriptString;
+    updateExpression();
+
+    Q_EMIT expressionChanged();
+    invalidate();
+}
+
+void ExpressionFilter::proxyModelCompleted(const QQmlSortFilterProxyModel& proxyModel)
+{
+    updateContext(proxyModel);
+}
+
+bool ExpressionFilter::filterRow(const QModelIndex& sourceIndex, const QQmlSortFilterProxyModel& proxyModel) const
+{
+    if (!m_scriptString.isEmpty()) {
+        QVariantMap modelMap;
+        QHash<int, QByteArray> roles = proxyModel.roleNames();
+
+        QQmlContext context(qmlContext(this));
+        auto addToContext = [&] (const QString &name, const QVariant& value) {
+            context.setContextProperty(name, value);
+            modelMap.insert(name, value);
+        };
+
+        for (auto it = roles.cbegin(); it != roles.cend(); ++it)
+            addToContext(it.value(), proxyModel.sourceData(sourceIndex, it.key()));
+        addToContext("index", sourceIndex.row());
+
+        context.setContextProperty("model", modelMap);
+
+        QQmlExpression expression(m_scriptString, &context);
+        QVariant variantResult = expression.evaluate();
+
+        if (expression.hasError()) {
+            qWarning() << expression.error();
+            return true;
+        }
+        if (variantResult.canConvert<bool>()) {
+            return variantResult.toBool();
+        } else {
+            qWarning("%s:%i:%i : Can't convert result to bool",
+                     expression.sourceFile().toUtf8().data(),
+                     expression.lineNumber(),
+                     expression.columnNumber());
+            return true;
+        }
+    }
+    return true;
+}
+
+void ExpressionFilter::updateContext(const QQmlSortFilterProxyModel& proxyModel)
+{
+    delete m_context;
+    m_context = new QQmlContext(qmlContext(this), this);
+    // what about roles changes ?
+    QVariantMap modelMap;
+
+    auto addToContext = [&] (const QString &name, const QVariant& value) {
+        m_context->setContextProperty(name, value);
+        modelMap.insert(name, value);
+    };
+
+    for (const QByteArray& roleName : proxyModel.roleNames().values())
+        addToContext(roleName, QVariant());
+
+    addToContext("index", -1);
+
+    m_context->setContextProperty("model", modelMap);
+    updateExpression();
+}
+
+void ExpressionFilter::updateExpression()
+{
+    if (!m_context)
+        return;
+
+    delete m_expression;
+    m_expression = new QQmlExpression(m_scriptString, m_context, 0, this);
+    connect(m_expression, &QQmlExpression::valueChanged, this, &ExpressionFilter::invalidate);
+    m_expression->setNotifyOnValueChanged(true);
+    m_expression->evaluate();
+}
+
+}
